+++ python/hugr_qir/cli.py
@@ -1,15 +1,22 @@
 """Cli for hugr-qir."""
 
+import logging
 import tempfile
 from importlib.metadata import version
 from pathlib import Path
-from typing import IO
 
 import click
 from quantinuum_qircheck import qircheck
 from quantinuum_qircheck.qircheck import ValidationError
 
-from hugr_qir._hugr_qir import cli
+from hugr_qir._hugr_qir import (
+    cli,
+    compile_target_choices,
+    opt_level_choices,
+)
+from hugr_qir.output import OutputFormat, get_write_mode, ir_string_to_output_format
+
+logger = logging.getLogger()
 
 
 @click.command(name="hugr-qir")
@@ -27,16 +34,46 @@ from hugr_qir._hugr_qir import cli
     help="Whether to validate the input hugr before and after each internal pass",
 )
 @click.option(
+    "-t",
+    "--target",
+    "target",
+    type=click.Choice(compile_target_choices()),
+    default=None,
+    help="LLVM compile target",
+)
+@click.option(
+    "-l",
+    "--opt-level",
+    "opt_level",
+    type=click.Choice(opt_level_choices()),
+    default=None,
+    help="LLVM optimization level",
+)
+@click.option(
+    "-f",
+    "--output-format",
+    "output_format",
+    type=click.Choice([c.value for c in OutputFormat], case_sensitive=False),
+    default="llvm-ir",
+    help="Choice of output format",
+)
+@click.option(
     "-o",
     "--output",
     "outfile",
-    type=click.File("w"),
-    default="-",
+    type=click.Path(path_type=Path),
+    default=None,
     help="Name of output file (optional)",
 )
 @click.version_option(version=version("hugr_qir"))
-def hugr_qir(
-    validate_qir: bool, validate_hugr: bool, hugr_file: Path, outfile: IO
+def hugr_qir(  # noqa: PLR0913
+    validate_qir: bool,
+    validate_hugr: bool,
+    target: str | None,
+    opt_level: str | None,
+    output_format: str,
+    hugr_file: Path,
+    outfile: Path | None,
 ) -> None:
     """Convert a HUGR file to QIR.
 
@@ -44,13 +81,37 @@ def hugr_qir(
     Per default, QIR is emitted to stdout, but can
     be written to a file using the `-o` option.
     """
-    hugr_qir_impl(validate_qir, validate_hugr, hugr_file, outfile)
+    hugr_qir_impl(
+        validate_qir,
+        validate_hugr,
+        target,
+        opt_level,
+        OutputFormat(output_format),
+        hugr_file,
+        outfile,
+    )
 
 
-def hugr_qir_impl(
-    validate_qir: bool, validate_hugr: bool, hugr_file: Path, outfile: IO
+def hugr_qir_impl(  # noqa: PLR0913
+    validate_qir: bool,
+    validate_hugr: bool,
+    target: str | None,
+    opt_level: str | None,
+    output_format: OutputFormat,
+    hugr_file: Path,
+    outfile: Path | None,
 ) -> None:
     options = ["-q"]
+    if target:
+        options.extend(["-t", target])
+    if opt_level:
+        options.extend(["-l", opt_level])
+        if opt_level == "none":
+            logger.warning(
+                "WARNING: Chosen optimization level"
+                " `none` will generally not result"
+                " in valid QIR."
+            )
     if validate_hugr:
         options.append("--validate")
     with tempfile.TemporaryDirectory(ignore_cleanup_errors=True) as tmp_dir:
@@ -67,7 +128,14 @@ def hugr_qir_impl(
             msg = f"Found a problem in the generated QIR: {e}"
             raise ValueError(msg) from e
 
-    outfile.write(qir)
+    llvm_write_mode = get_write_mode(output_format, outfile)
+    qir_out = ir_string_to_output_format(qir, output_format)
+
+    if outfile:
+        with outfile.open(mode=llvm_write_mode) as output:
+            output.write(qir_out)
+    else:
+        print(qir_out)
 
 
 if __name__ == "__main__":